fix: entire auth flow
