+++ longhorn-studies/app/(app)/(tabs)/home/index.tsx
@@ -1,18 +1,19 @@
-import { Text, View } from 'react-native';
+import { View } from 'react-native';
 
+import { Button } from '~/components/Button';
 import { useAuth } from '~/store/AuthProvider';
 
-export default function Index() {
+export default function Home() {
   const { signOut } = useAuth();
   return (
     <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
-      <Text
+      <Button
+        title="Sign Out"
         onPress={() => {
           // The `app/(app)/_layout.tsx` will redirect to the sign-in screen.
           signOut();
-        }}>
-        Sign Out
-      </Text>
+        }}
+      />
     </View>
   );
 }
